backend-doc-reader/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
python-docx
python-dotenv
sentence-transformers
transformers
torch